Implemented all controllers Version 1.0 working

+++ src/main/kotlin/com/example/guera/DataProvisioner/Services/BoardService.kt
@@ -43,7 +43,7 @@ class BoardService(
 
     override fun remove(id: UUID): Boolean {
         val exists = boardRepository.existsById(id)
-        boardRepository.deleteById(id)
+        if (exists) boardRepository.deleteById(id)
         return exists
     }
 }
